pyproject.toml
[tool.poetry]
name = "es_outfitter"
version = "0.1.0"
description = "A web-based application to customize outfit configurations for ships in the space game 'Endless Sky'"
authors = ["thorndeux <christof.koegler@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
